combine char-budget and skill-loader reorganize branching add unit tests
